make test builder available to all tests, add example for script offset macro.
