+++ tests/outputs.rs
@@ -1,3 +1,4 @@
+use fuel_vm::util::test_helpers::TestBuilder;
 use fuel_vm::{
     consts::{REG_ONE, REG_ZERO},
     prelude::*,
@@ -276,162 +277,3 @@ fn variable_output_increased_by_contract_transfer_out() {}
 #[test]
 #[ignore]
 fn variable_output_not_increased_by_contract_transfer_out_on_revert() {}
-
-pub struct TestBuilder {
-    rng: StdRng,
-    gas_price: Word,
-    gas_limit: Word,
-    byte_price: Word,
-    inputs: Vec<Input>,
-    outputs: Vec<Output>,
-    script: Vec<Opcode>,
-    script_data: Vec<u8>,
-    witness: Vec<Witness>,
-    storage: MemoryStorage,
-}
-
-impl TestBuilder {
-    pub fn new(seed: u64) -> Self {
-        TestBuilder {
-            rng: StdRng::seed_from_u64(seed),
-            gas_price: 0,
-            gas_limit: 100,
-            byte_price: 0,
-            inputs: Default::default(),
-            outputs: Default::default(),
-            script: vec![Opcode::RET(REG_ONE)],
-            script_data: vec![],
-            witness: vec![Witness::default()],
-            storage: MemoryStorage::default(),
-        }
-    }
-
-    pub fn gas_price(&mut self, price: Word) -> &mut TestBuilder {
-        self.gas_price = price;
-        self
-    }
-
-    pub fn gas_limit(&mut self, limit: Word) -> &mut TestBuilder {
-        self.gas_limit = limit;
-        self
-    }
-
-    pub fn byte_price(&mut self, price: Word) -> &mut TestBuilder {
-        self.byte_price = price;
-        self
-    }
-
-    pub fn change_output(&mut self, color: Color) -> &mut TestBuilder {
-        self.outputs.push(Output::change(self.rng.gen(), 0, color));
-        self
-    }
-
-    pub fn coin_output(&mut self, color: Color, amount: Word) -> &mut TestBuilder {
-        self.outputs.push(Output::coin(self.rng.gen(), amount, color));
-        self
-    }
-
-    pub fn withdrawal_output(&mut self, color: Color, amount: Word) -> &mut TestBuilder {
-        self.outputs.push(Output::withdrawal(self.rng.gen(), amount, color));
-        self
-    }
-
-    pub fn variable_output(&mut self, color: Color) -> &mut TestBuilder {
-        self.outputs.push(Output::variable(self.rng.gen(), 0, color));
-        self
-    }
-
-    pub fn contract_output(&mut self, id: &ContractId) -> &mut TestBuilder {
-        let input_idx = self
-            .inputs
-            .iter()
-            .find_position(|input| matches!(input, Input::Contract {contract_id, ..} if contract_id == id))
-            .expect("expected contract input with matching contract id");
-        self.outputs
-            .push(Output::contract(input_idx.0 as u8, self.rng.gen(), self.rng.gen()));
-        self
-    }
-
-    pub fn coin_input(&mut self, color: Color, amount: Word) -> &mut TestBuilder {
-        self.inputs.push(Input::coin(
-            self.rng.gen(),
-            self.rng.gen(),
-            amount,
-            color,
-            0,
-            0,
-            vec![],
-            vec![],
-        ));
-        self
-    }
-
-    pub fn contract_input(&mut self, contract_id: ContractId) -> &mut TestBuilder {
-        self.inputs.push(Input::contract(
-            self.rng.gen(),
-            self.rng.gen(),
-            self.rng.gen(),
-            contract_id,
-        ));
-        self
-    }
-
-    pub fn script(&mut self, script: Vec<Opcode>) -> &mut TestBuilder {
-        self.script = script;
-        self
-    }
-
-    pub fn script_data(&mut self, script_data: Vec<u8>) -> &mut TestBuilder {
-        self.script_data = script_data;
-        self
-    }
-
-    pub fn witness(&mut self, witness: Vec<Witness>) -> &mut TestBuilder {
-        self.witness = witness;
-        self
-    }
-
-    pub fn storage(&mut self, storage: MemoryStorage) -> &mut TestBuilder {
-        self.storage = storage;
-        self
-    }
-
-    pub fn build(&mut self) -> Transaction {
-        Transaction::script(
-            self.gas_price,
-            self.gas_limit,
-            self.byte_price,
-            0,
-            self.script.iter().copied().collect(),
-            self.script_data.clone(),
-            self.inputs.clone(),
-            self.outputs.clone(),
-            self.witness.clone(),
-        )
-    }
-
-    pub fn execute(&mut self) -> Vec<Output> {
-        let tx = self.build();
-        let mut client = MemoryClient::new(self.storage.clone());
-        client.transact(tx);
-        let txtor: Transactor<_> = client.into();
-        let outputs = txtor.state_transition().unwrap().tx().outputs();
-        outputs.to_vec()
-    }
-
-    pub fn execute_get_change(&mut self, find_color: Color) -> Word {
-        let outputs = self.execute();
-        let change = outputs.into_iter().find_map(|output| {
-            if let Output::Change { amount, color, .. } = output {
-                if &color == &find_color {
-                    Some(amount)
-                } else {
-                    None
-                }
-            } else {
-                None
-            }
-        });
-        change.expect(format!("no change matching color {:x} was found", &find_color).as_str())
-    }
-}
